detection-violence-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
bcrypt==3.2.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.68.1
h11==0.16.0
idna==3.10
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==1.10.21
pydantic_core==2.33.1
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.23
starlette==0.14.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.15.0
moviepy
